Add function to check if DOM has changed, and improved clickElement()
clickElement() now takes an argument that specifies if clicking the element triggers a change in the dom or navigating to another url. It now does not rely on a set interval for the callback, but rather checks to see if the url has changed or if the DOM has changed (using Levenshtein distance) depending on the expectAjax argument.